rolling through to solver and runner
