+++ tools/mudly/internal/config/config.go
@@ -1,34 +1,41 @@
 package config
 
-import (
-	"ponglehub.co.uk/tools/mudly/internal/steps"
-	"ponglehub.co.uk/tools/mudly/internal/target"
-)
+import "ponglehub.co.uk/tools/mudly/internal/target"
 
-type Runnable interface {
-	Run(dir string, artefact string, env map[string]string) steps.CommandResult
-}
-
-type DevEnv struct {
-	Compose *map[string]interface{} `yaml:"compose"`
-}
-
-type Pipeline struct {
-	Name  string            `yaml:"name"`
-	Steps []Runnable        `yaml:"-"`
-	Env   map[string]string `yaml:"env"`
-}
-
-type Artefact struct {
-	Name         string            `yaml:"name"`
-	Pipeline     Pipeline          `yaml:"pipeline"`
-	Dependencies []target.Target   `yaml:"dependencies"`
-	Env          map[string]string `yaml:"env"`
+type Dockerfile struct {
+	Name    string
+	Content string
 }
 
 type Config struct {
-	DevEnv    *DevEnv
-	Path      string
-	Artefacts []Artefact
-	Env       map[string]string `yaml:"env"`
+	Path       string
+	Dockerfile []Dockerfile
+	Artefacts  []Artefact
+	Pipelines  []Pipeline
+	Env        map[string]string
+}
+
+type Pipeline struct {
+	Name  string
+	Env   map[string]string
+	Steps []Step
+}
+
+type Artefact struct {
+	Name      string
+	DependsOn []target.Target
+	Env       map[string]string
+	Steps     []Step
+	Pipeline  string
+}
+
+type Step struct {
+	Name       string
+	Env        map[string]string
+	Condition  string
+	Command    string
+	Watch      []string
+	Dockerfile string
+	Context    string
+	Tag        string
 }
