+++ tools/mudly/internal/config/fs.go
@@ -1,4 +1,4 @@
-package config_new
+package config
 
 import "io/ioutil"
