+++ tools/mudly/internal/config/reader_test.go
@@ -1,4 +1,4 @@
-package config_new
+package config
 
 import (
 	"testing"
