+++ tools/mudly/internal/runner/runner.go
@@ -4,16 +4,14 @@ import (
 	"fmt"
 
 	"github.com/sirupsen/logrus"
-	"ponglehub.co.uk/tools/mudly/internal/solver"
-	"ponglehub.co.uk/tools/mudly/internal/steps"
 )
 
 type runResult struct {
-	node    *solver.Node
+	node    *Node
 	success bool
 }
 
-func Run(nodes []*solver.Node) (err error) {
+func Run(nodes []*Node) (err error) {
 	numRunning := 0
 	outputChan := make(chan runResult, 10)
 
@@ -22,7 +20,7 @@ func Run(nodes []*solver.Node) (err error) {
 
 		for _, node := range pending {
 			numRunning += 1
-			node.State = solver.STATE_RUNNING
+			node.State = STATE_RUNNING
 			go runNode(node, outputChan)
 		}
 
@@ -41,14 +39,14 @@ func Run(nodes []*solver.Node) (err error) {
 	}
 }
 
-func getRunnableNodes(nodes []*solver.Node) []*solver.Node {
-	runnables := []*solver.Node{}
+func getRunnableNodes(nodes []*Node) []*Node {
+	runnables := []*Node{}
 
 	for _, node := range nodes {
-		runnable := node.State == solver.STATE_PENDING
+		runnable := node.State == STATE_PENDING
 
 		for _, dep := range node.DependsOn {
-			if dep.State != solver.STATE_COMPLETE && dep.State != solver.STATE_SKIPPED {
+			if dep.State != STATE_COMPLETE && dep.State != STATE_SKIPPED {
 				runnable = false
 			}
 		}
@@ -61,11 +59,11 @@ func getRunnableNodes(nodes []*solver.Node) []*solver.Node {
 	return runnables
 }
 
-func depsSkipped(node *solver.Node) bool {
+func depsSkipped(node *Node) bool {
 	skipped := node.DependsOn != nil && len(node.DependsOn) > 0
 
 	for _, dep := range node.DependsOn {
-		if dep.State != solver.STATE_SKIPPED {
+		if dep.State != STATE_SKIPPED {
 			skipped = false
 		}
 	}
@@ -73,7 +71,7 @@ func depsSkipped(node *solver.Node) bool {
 	return skipped
 }
 
-func runNode(node *solver.Node, outputChan chan<- runResult) {
+func runNode(node *Node, outputChan chan<- runResult) {
 	if depsSkipped(node) {
 		node.SharedEnv["DEPS_SKIPPED"] = "true"
 	}
@@ -83,12 +81,12 @@ func runNode(node *solver.Node, outputChan chan<- runResult) {
 	success := true
 
 	switch result {
-	case steps.COMMAND_SUCCESS:
-		node.State = solver.STATE_COMPLETE
-	case steps.COMMAND_SKIPPED:
-		node.State = solver.STATE_SKIPPED
-	case steps.COMMAND_ERROR:
-		node.State = solver.STATE_ERROR
+	case COMMAND_SUCCESS:
+		node.State = STATE_COMPLETE
+	case COMMAND_SKIPPED:
+		node.State = STATE_SKIPPED
+	case COMMAND_ERROR:
+		node.State = STATE_ERROR
 		success = false
 	}
