--- tools/mudly/internal/steps/status.go
@@ -1,9 +1,0 @@
-package steps
-
-type CommandResult int
-
-const (
-	COMMAND_SUCCESS CommandResult = iota
-	COMMAND_ERROR
-	COMMAND_SKIPPED
-)
